Переношу функции непосредственно связанные с событиями, в отдельный файл, сделала выбор городов в форме редактирования
