+++ src/components/event-card.js
@@ -2,33 +2,6 @@ import AbstractComponent from './abstract-component';
 import {createEventCardTemplate} from './templates/event-card';
 import {formatDate} from './templates/date';
 
-const EventTypeTransport = {
-  TAXI: `taxi`,
-  BUS: `bus`,
-  TRAIN: `train`,
-  SHIP: `ship`,
-  TRANSPORT: `transport`,
-  DRIVE: `drive`,
-  FLIGHT: `flight`,
-};
-
-const EventTypePlace = {
-  CHECKIN: `check-in`,
-  SIGHTSEEING: `sightseeing`,
-  RESTAURANT: `restaurant`
-};
-
-const Preposition = {
-  TRANSPORT: `to`,
-  PLACE: `in`,
-};
-
-const getCorrectPreposition = (type) => {
-  return ((type === EventTypePlace.CHECKIN || type === EventTypePlace.SIGHTSEEING || type === EventTypePlace.RESTAURANT) ?
-    Preposition.PLACE : Preposition.TRANSPORT);
-};
-
-const createPhotos = (destination) => Object.values(destination.pictures);
 
 class EventCardComponent extends AbstractComponent {
   constructor(events) {
@@ -51,8 +24,4 @@ class EventCardComponent extends AbstractComponent {
 }
 
 export default EventCardComponent;
-export {EventTypeTransport,
-  EventTypePlace,
-  createPhotos,
-  getCorrectPreposition,
-};
+
