+++ src/components/trip-info.js
@@ -2,15 +2,6 @@
 import AbstractComponent from './abstract-component';
 import {createTripInfoTemplate} from './templates/trip-info';
 
-const generateEventsWithUniqueDestinationNames = (events) => {
-  const eventsFiltered = events.filter((eventItem, index, self) =>
-    index === self.findIndex((t) => (
-      t.destination.name === eventItem.destination.name
-    ))
-  );
-  return eventsFiltered;
-};
-
 class TripInfoComponent extends AbstractComponent {
   constructor(events) {
     super();
@@ -24,4 +15,4 @@ class TripInfoComponent extends AbstractComponent {
 }
 
 export default TripInfoComponent;
-export {generateEventsWithUniqueDestinationNames};
+
